Resync with DelayedArray 0.35.2
nzindex() got removed from DelayedArray 0.35.2 (BioC 3.22) after
being deprecated in BioC 3.20 and defunct in BioC 3.21. The
replacement is to use SparseArray::nzwhich(x, arr.ind=TRUE)

Also use nzvals() instead of nzdata(). The latter is just a slot accessor
that only works on COO_SparseArray objects while the former works on **any**
SparseArray derivative (i.e. COO_SparseArray and SVT_SparseArray objects).
It even works on dgCMatrix objects from the Matrix package.
Furthermore, nzdata() is not guaranteed to return the non-zero values in
an order that is consistent with the index returned by nzwhich(), but
nzvals() is. See '?nzwhich' in the SparseArray package.